use bcrypt package
This replaces python-bcrypt with bcrypt. The latter is more secure by
default, using `2b` hashes. A test which formerly passed, taking in an
integer instead of a string, has been removed as this behavior is no
longer supported: always supply a string as it is the only legal type.

fixes #30