Add the `BCRYPT_HANDLE_LONG_PASSWORDS` configuration value
By default, the bcrypt algorithm has a maximum password length of 72 bytes
and ignores any bytes beyond that. A common workaround is to hash the
given password using a cryptographic hash (such as `sha256`), take its
hexdigest to prevent NULL byte problems, and hash the result with bcrypt.
If the `BCRYPT_HANDLE_LONG_PASSWORDS` configuration value is set to `True`,
the workaround described above will be enabled.

WARNING: enabling this option on a project that is already using Flask-Bcrypt
will break password checking.